Move Section Block Factories to the core app's tests
